Created a fucnntion to build dependencies for a given module. This can be used to pull out all of the packages needed to run the generated code.
For the UD, DW, and FG modules the generated code for a given module output will now have the needed library calls appended at the top.

When saving the module state the app code generated will have the correct library calls needed at the top of the script that's generated.